+++ packages/pluggableWidgets/file-uploader-web/CHANGELOG.md
@@ -10,6 +10,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - We made "Associated files/images" preconfigured with corresponding entities.
 
+- We made it possible to configure custom buttons for file uploader entries.
+
 ### Fixed
 
 - We fixed an issue with hover colors of the dropzone